requirements.txt
fpdf2==2.7.8
pytest==8.0.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
